move unitValue function into units module, and re-export from further up tree
