+++ mapping/dwellingFabricMapper.test.ts
@@ -10,7 +10,7 @@ import { mapCeilingAndRoofData, mapDoorData, mapFloorData, mapLightingData, mapT
 import { defaultZoneName } from "./common";
 import type { DwellingSpaceLightingData, DwellingSpaceZoneParametersData } from "~/stores/ecaasStore.schema";
 import { centimetre, millimetre } from "../utils/units/length";
-import { unitValue } from "~/utils/units/types";
+import { unitValue } from "~/utils/units";
 
 type BuildingElementGround = Extract<SchemaBuildingElement, { type: "BuildingElementGround" }>;
 type BuildingElementOpaque = Extract<SchemaBuildingElement, { type: "BuildingElementOpaque" }>;
