+++ pages/dwelling-fabric/floors/ground/ground-floor.test.ts
@@ -4,7 +4,7 @@ import { screen } from "@testing-library/vue";
 import GroundFloor from "./[floor].vue";
 import { FloorType, MassDistributionClass, WindShieldLocation } from "~/schema/api-schema.types";
 import { centimetre } from "~/utils/units/length";
-import { unitValue } from "~/utils/units/types";
+import { unitValue } from "~/utils/units";
 
 const navigateToMock = vi.hoisted(() => vi.fn());
 mockNuxtImport("navigateTo", () => {
