+++ pages/infiltration-and-ventilation/mechanical-ventilation/[mechanical].vue
@@ -3,7 +3,7 @@ import { v4 as uuidv4 } from "uuid";
 import type { MVHRLocation } from "~/schema/api-schema.types";
 import { VentType } from "~/schema/api-schema.types";
 import { litrePerSecond } from "~/utils/units/flowRate";
-import { unitValue } from "~/utils/units/types";
+import { unitValue } from "~/utils/units";
 import { getUrl } from "#imports";
 
 const title = "Mechanical ventilation";
