+++ pages/infiltration-and-ventilation/mechanical-ventilation/mechanical-form.test.ts
@@ -5,7 +5,7 @@ import { screen } from "@testing-library/vue";
 import { v4 as uuidv4 } from "uuid";
 import { MVHRLocation, VentType } from "~/schema/api-schema.types";
 import { litrePerSecond } from "~/utils/units/flowRate";
-import { unitValue } from "~/utils/units/types";
+import { unitValue } from "~/utils/units";
 
 describe("mechanical ventilation form", () => {
 	const user = userEvent.setup();
