+++ utils/units/types.ts
@@ -16,13 +16,3 @@ export type UnitValue = {
 	amount: number,
 	unit: UnitName
 };
-
-export function unitValue<T extends Unit>(amount: number, unit: T): Dimension<T["name"]>;
-export function unitValue<T extends UnitName>(amount: number, unit: T): Dimension<T>;
-export function unitValue(amount: number, unit: Unit | UnitName) {
-	const unitName = typeof unit !== "string" ? unit.name : unit;
-	return {
-		amount,
-		unit: unitName,
-	};
-}
